requirements.txt
pytest
selene>=2.0.0rc4
strenum
allure-pytest==2.9.45